Character Pruning
Adds 2 options for character pruning: Minimum (the current functionality) and Extreme (which removes all characters except those closely related to the ruler). Additionally, there is functional code for a remove minor characters setting, but for the time being, I have decided to not enable it as an option, as removing minor characters only makes a negligible difference. Finally, the BADirectory has been removed from Fronter, as the converter now relies solely on the save file for the BA mod location.