Automatically Grant City Status for Metropoli
Previously, there were a few edge-cases where it technically was possible for a huge BA metropolis to become an I:R territory if enough BA territories were in the combined I:R province. Now, city status is automatically granted to an I:R province with at least 5 pops and had a BA metropolis.