Pop Total Adjustment and More ExoMappings
Adjusts the total pops in the converted scope from 6,000 to 8,580 to be more in-line with the actual total pops from Invictus/vanilla. Additionally, the globalPopTotal has been moved into main from the method a couple of new exomappings have a tag defined, primarily because of their importance in mission trees.