Divorce Long Distance Relationships
Female characters defined in setup always move to the country of their husbands, including when the female is the ruler of a country. When this happens, the game auto-generates a new character with the ID after the husband's, causing all other character ID's to be out-of-sync by 1. The converter will now automatically divorce couples living in two different countries to prevent this from happening.